Persis input value
