+++ src/pages/HomePage.jsx
@@ -9,6 +9,7 @@ import {
   MEGA_SENA_VALUE,
 } from "../constants";
 import { binomial } from "../math.functions";
+import { Options } from "../components/Options";
 
 export const HomePage = () => {
   const [value, setValue] = useState(MEGA_SENA_VALUE.toFixed(2));
@@ -18,7 +19,6 @@ export const HomePage = () => {
     isValid: true,
     message: "",
   });
-  const [showOptions, setShowOptions] = useState(false);
 
   const handleGenerate = (evt) => {
     evt.preventDefault();
@@ -83,81 +83,6 @@ export const HomePage = () => {
     setBets(newBets);
   };
 
-  const handleFixedInputChange = (value) => {
-    setFixedInput(value);
-
-    if (!/[0-9,.]+/.test(value)) {
-      setFixedInput((old) => ({
-        ...old,
-        isValid: false,
-        // We also silently accept '.', but whatever
-        message: "Apenas números e vírgulas são aceitos",
-        numbers: [],
-      }));
-      return;
-    }
-
-    const trimmed = value
-      .replace(".", ",")
-      .replace(/\s/g, "")
-      .replace(/,+$/, "") // Remove , from the end
-      .replace(/^,+/, ""); // Remove , from the beggining
-
-    if (trimmed.length === 0) {
-      setFixedInput((old) => ({
-        ...old,
-        isValid: true,
-        message: "",
-        numbers: [],
-      }));
-      return;
-    }
-
-    const numbers = trimmed.split(",").map(Number);
-
-    if (numbers.length > 5) {
-      setFixedInput((old) => ({
-        ...old,
-        numbers: [],
-        isValid: false,
-        message: "Não é possível fixar mais do que 5 números",
-      }));
-      return;
-    }
-
-    const betweenLimits = numbers
-      .map((n) => n <= MEGA_SENA_MAX && n >= MEGA_SENA_MIN)
-      .reduce((a, b) => a && b);
-
-    if (!betweenLimits) {
-      setFixedInput((old) => ({
-        ...old,
-        numbers: [],
-        isValid: false,
-        message: `Os valores devem estar entre ${MEGA_SENA_MIN} e ${MEGA_SENA_MAX}`,
-      }));
-      return;
-    }
-
-    const setNumbers = new Set(numbers);
-    if (setNumbers.size !== numbers.length) {
-      setFixedInput((old) => ({
-        ...old,
-        numbers: [],
-        isValid: false,
-        message: "Existem valores repetidos",
-      }));
-      return;
-    }
-
-    setFixedInput((old) => ({
-      ...old,
-      isValid: true,
-      message: "",
-      numbers,
-    }));
-  };
-
   return (
     <div>
       <form className="container">
@@ -192,36 +117,7 @@ export const HomePage = () => {
         </div>
 
         <BetsTable bets={bets} />
-
-        <div className="row" align="center">
-          <div className="col-12">
-            <button
-              onClick={() => setShowOptions((o) => !o)}
-              className="btn btn-secondary"
-              type="button"
-            >
-              Opções
-            </button>
-          </div>
-        </div>
-
-        {showOptions && (
-          <div
-            className={
-              fixedInput.isValid ? "alert m-1" : "alert alert-danger m-1"
-            }
-          >
-            Incluir números fixos para aparecer nas apostas?{" "}
-            <input
-              id="fixed-input"
-              placeholder="1, 2, 3"
-              onChange={(e) => handleFixedInputChange(e.target.value)}
-            />
-            {fixedInput.message && (
-              <p className="alert">{fixedInput.message}</p>
-            )}
-          </div>
-        )}
+        <Options fixedInput={fixedInput} setFixedInput={setFixedInput} />
       </form>
     </div>
   );
